re arranged graphql queries and mutations into respective .py files. Included graphql schema in setup.py
