+++ backend/.gitignore
@@ -1,6 +1,6 @@
 .env
 .pytest_cache
-.idea
+.idea/
 
 ### Flask ###
 instance/*
