+++ backend/src/app/__init__.py
@@ -2,4 +2,5 @@ __version__ = '0.1.0'
 
 from .functions import *
 from .setup import createApp
-from .schema import schema
+# from .schema import schema
+
